New structure applied: Services, ErrorHandler, Storages and Storage Interfaces. Added new fields for User - friends and Film - likes.
