+++ src/main/java/ru/yandex/practicum/filmorate/controller/ErrorHandler.java
@@ -1,0 +1,5 @@
+package ru.yandex.practicum.filmorate.controller;
+
+public class ErrorHandler {
+
+}
