+++ src/main/java/ru/yandex/practicum/filmorate/model/ErrorResponse.java
@@ -1,0 +1,15 @@
+package ru.yandex.practicum.filmorate.model;
+
+public class ErrorResponse {
+    // название ошибки
+    String error;
+
+    public ErrorResponse(String error) {
+        this.error = error;
+    }
+
+    // геттеры необходимы, чтобы Spring Boot мог получить значения полей
+    public String getError() {
+        return error;
+    }
+}
