+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 import org.hibernate.validator.constraints.Length;
 
 import java.time.LocalDate;
+import java.util.Set;
 
 /**
  * Film.
@@ -16,6 +17,7 @@ import java.time.LocalDate;
 @EqualsAndHashCode
 public class Film {
     private Long id;
+    private Set<Long> likes;
 
     @NotBlank
     private String name;
