+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -6,12 +6,14 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.time.LocalDate;
+import java.util.Set;
 
 @Getter
 @Setter
 @EqualsAndHashCode(of = {"email"})
 public class User {
     private Long id;
+    private Set<Long> friends;
 
     @NotBlank
     private String email;
