+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -1,0 +1,16 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.util.Collection;
+import java.util.Map;
+
+public interface FilmStorage {
+    public Collection<Film> findAll();
+
+    public Film create(Film film);
+
+    public Film update(Film newFilm);
+
+    public Map<Long, Film> getFilms();
+}
